speed up obtaining VCS branch.
Running actual VCS command could be extremely slow. Fortunately, we
don't have to run `hg branch` command unless the `.hg` directory exists
either in the current directory or in any of the ancestor directories.

Thus if we check for the VCS-specific directory in advance we could avoid
running the command unless the directory exists.